+++ assets/js/gmaps.js
@@ -4,14 +4,23 @@ function initialize() {
     var lat = jQuery(this).attr('data-lat');
     var lng = jQuery(this).attr('data-lng');
     var zoom = parseInt(jQuery(this).attr('data-zoom'));
+    var disable_zoom = (jQuery(this).attr('data-disable-zoom').toLowerCase() === 'true')
+    var disable_draggable = (jQuery(this).attr('data-disable-draggable').toLowerCase() === 'true')
     var kml = jQuery(this).attr('data-kml');
     var marker = jQuery(this).attr('data-marker');
     
     var myLatLng = new google.maps.LatLng(lat, lng);
-    var mapOptions = {
-      zoom: zoom,
-      center: myLatLng
-    };
+    var mapOptions = new google.maps.MapOptions();
+    mapOptions.zoom = zoom;
+    mapOptions.center = myLatLng;
+    if (disable_zoom){
+      mapOptions.draggable = false;
+    }
+    if (disable_zoom) {
+      mapOptions.zoomControl = false;
+      mapOptions.scrollwheel = false;
+      mapOptions.disableDoubleClickZoom = true;
+    }
   
     var map = new google.maps.Map(jQuery(this)[0], mapOptions);