+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.3
+
+- Add options to deactivate zoom and draggable function
+
 # 0.2
 
 - Add KML layer support